+++ settings/base.py
@@ -147,6 +147,7 @@ PREREQ_APPS = [
     "django.contrib.auth",
     "django.contrib.contenttypes",
     "django.contrib.sessions",
+    "django.contrib.sitemaps",
     "django.contrib.sites",
     "django.contrib.messages",
     "django.contrib.humanize",