+++ tests/test_views.py
@@ -36,10 +36,40 @@ def test_robots_txt(db, tp):
     assert response.status_code == 200
 
 
-def test_sitemap(db, tp):
-    url = tp.reverse("sitemap")
+def test_sitemap_index(db, tp):
+    url = tp.reverse("sitemap_index")
     response = tp.client.get(url)
     assert response.status_code == 200
+
+
+def test_sitemap_static_section(db, tp):
+    url = tp.reverse("sitemaps", section="static")
+    response = tp.client.get(url)
+    assert response.status_code == 200
+
+
+def test_sitemap_packages_section(db, tp):
+    url = tp.reverse("sitemaps", section="packages")
+    response = tp.client.get(url)
+    assert response.status_code == 200
+
+
+def test_sitemap_grids_section(db, tp):
+    url = tp.reverse("sitemaps", section="grids")
+    response = tp.client.get(url)
+    assert response.status_code == 200
+
+
+def test_sitemap_blog_section(db, tp):
+    url = tp.reverse("sitemaps", section="blog")
+    response = tp.client.get(url)
+    assert response.status_code == 200
+
+
+def test_sitemap_invalid_section(db, tp):
+    url = tp.reverse("sitemaps", section="invalid")
+    response = tp.client.get(url)
+    assert response.status_code == 404
 
 
 def test_syndication(db, tp):